testing

+++ src/base/TensorFFT256.cu
@@ -64,8 +64,8 @@ __global__ void TensorFFT256(__half* input_data_RE, __half* input_data_IM,
     //            static_cast<__half>(8.0));
     float phase = (static_cast<float>(j * inter_warp_id_16) * M_PI) / 8.0;
 
-    buffer_RE[buffer_array_id] = cosf(phase);
-    buffer_IM[buffer_array_id] = -sinf(phase);
+    buffer_RE[buffer_array_id] = cos(phase);
+    buffer_IM[buffer_array_id] = -sin(phase);
   }
 
   //Literal version of dft matrix.
@@ -237,8 +237,8 @@ __global__ void TensorFFT256(__half* input_data_RE, __half* input_data_IM,
     //            static_cast<__half>(128.0));
     float phase = (static_cast<float>(inter_warp_id_16 * j) * M_PI) / 128.0;
 
-    __half twiddle_RE = cosf(phase);
-    __half twiddle_IM = -sinf(phase);
+    __half twiddle_RE = cos(phase);
+    __half twiddle_IM = -sin(phase);
 
     __half input_RE = buffer_RE[buffer_array_id];
     __half input_IM = buffer_IM[buffer_array_id];
